% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IC1}
\alias{IC1}
\title{sample interference dataset}
\format{IC1 is a data frame with 10000 cases (rows) and 6 variables
(columns) named thickness, biref, R, G, B and retardation.}
\usage{
IC1
}
\description{
Low-resolution computed interference data
}
\details{
Low-resolution data that would be obtained by running the
\code{InterferenceTable} function with
thickVect = \code{seq(0.01,50,length.out = 50)} micrometers  and
birefVect = \code{seq(2e-4,1e-2,by = 5e-4)}
}
\author{
Olivier Eterradossi, \email{olivier.eterradossi@mines-ales.fr}
}
\keyword{datasets}
