% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L.R
\name{L}
\alias{L}
\title{Computes the degree of transmission under crossed polarizer and analyzer}
\usage{
L(lambda = 550, d = 30, biref)
}
\arguments{
\item{lambda}{Wavelength (nanometers).}

\item{d}{Thickness (micrometers).}

\item{biref}{Birefringence .}
}
\value{
a single transmission value
}
\description{
\code{L} is called internally to fill the spectral transmission matrix
needed for interference color calculation
}
\examples{
\dontrun{
test.L<-L(lambda=550,d=30,biref=0.00025)
}
}
\references{
Bloss, F.D. (1999) Optical Cristallography. Mineralogical Society of America
Monograph Series, Washington DC, Publication #5. ISBN 0-939950-49-9
}
\author{
Olivier Eterradossi, \email{olivier.eterradossi@mines-ales.fr}
}
